Started refactoring GUI using MVC design pattern
